pyproject.toml
[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

# ── Project metadata ──────────────────────────────────────────
[project]
name        = "aura-autonomy-engine"
version     = "6.0.0"               # keep in sync with core/version.py
description = "Aura Sovereign Autonomous Cognitive Agent"
readme      = "README.md"
license     = { text = "Proprietary" }
authors     = [{ name = "Bryan", email = "" }]
requires-python = ">=3.9"

# Minimal runtime deps — extras handle the heavy stuff
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "httpx>=0.27",
    "aiofiles>=23.0",
    "python-multipart>=0.0.9",
    "PyYAML>=6.0",
    "psutil>=5.9",
    "ollama>=0.3",
]

[project.optional-dependencies]
# pip install "aura-autonomy-engine[ml]"
ml = [
    "torch>=2.3",
    "diffusers>=0.27",
    "transformers>=4.41",
    "chromadb>=0.5",
    "faiss-cpu>=1.8",
    "numpy>=1.26",
    "scipy>=1.13",
    "mlx>=0.14 ; sys_platform == 'darwin'",
    "mlx-lm>=0.4 ; sys_platform == 'darwin'",
]

# pip install "aura-autonomy-engine[voice]"
voice = [
    "sounddevice>=0.4",
    "pyaudio>=0.2",
    "pyttsx3>=2.90",
    "faster-whisper>=1.0",
    "playwright>=1.44",
]

# pip install "aura-autonomy-engine[senses]"
senses = [
    "playwright>=1.44",
    "pyautogui>=0.9",
    "pillow>=10.0",
    "python-nmap>=0.7",
    "netifaces>=0.11",
]

# pip install "aura-autonomy-engine[dev]"
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "black>=24.0",
    "ruff>=0.4",
    "pre-commit>=3.7",
    "httpx>=0.27",          # for TestClient
    "faker>=25.0",
]

# pip install "aura-autonomy-engine[all]"
all = [
    "aura-autonomy-engine[ml,voice,senses,dev]",
]

[project.scripts]
aura         = "aura_launcher:main"
aura-server  = "interface.server:main"
aura-bundle  = "utils.bundler:main"

# ── Tool: setuptools ──────────────────────────────────────────
[tool.setuptools.packages.find]
where   = ["."]
exclude = ["tests*", "scripts*", "docs*", "data*"]

# ── Tool: pytest ──────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode   = "auto"
testpaths      = ["tests"]
log_cli        = true
log_cli_level  = "WARNING"
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:pkg_resources",
    "ignore::PendingDeprecationWarning",
]
addopts = "--tb=short -q"

# ── Tool: mypy ────────────────────────────────────────────────
[tool.mypy]
python_version         = "3.11"
strict                 = false
ignore_missing_imports = true
warn_unused_ignores    = true
warn_return_any        = false
disallow_untyped_defs  = false   # too much legacy code; enable per-module
exclude                = ["scripts/", "scratch/", "dist/"]

[[tool.mypy.overrides]]
module = ["core.*", "infrastructure.*", "skills.*"]
ignore_missing_imports = true

# ── Tool: ruff (linter) ───────────────────────────────────────
[tool.ruff]
line-length    = 100
target-version = "py311"
src            = ["."]
exclude        = ["dist", "build", "scripts", "scratch", ".venv"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "RUF", # ruff-specific
]
ignore = [
    "E501",  # line too long (black handles this)
    "B008",  # do not perform function calls in default args (FastAPI pattern)
    "B904",  # raise from None in except blocks (intentional in some places)
]

[tool.ruff.lint.per-file-ignores]
"tests/*"  = ["S101"]   # assert is fine in tests
"scripts/*" = ["T201"]  # print() is fine in scripts

# ── Tool: black ───────────────────────────────────────────────
[tool.black]
line-length    = 100
target-version = ["py311"]
exclude        = "/(dist|build|\\.venv|scratch)/"

# ── Tool: coverage ────────────────────────────────────────────
[tool.coverage.run]
source   = ["core", "skills", "infrastructure", "interface"]
omit     = ["*/tests/*", "*/scripts/*", "*/scratch/*"]

[tool.coverage.report]
show_missing   = true
skip_covered   = false
fail_under     = 40     # raise as the codebase gets more tests
